refactor(validation): move print methods to TestRunner trait defaults
- Add default implementations for 6 print methods on the TestRunner trait, each delegating to the corresponding ValidationConsole method
- Remove boilerplate print delegation from MultiplicationRunner, ChainMultiplicationRunner, and ConvolutionRunner
